pyproject.toml
[tool.poetry]
name = "convert pdf to image"
version = "0.1.0"
description = ""
authors = ["hajja <mohamedhajja4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pdf2image = "1.16.3"
pillow = "10.1.0"
streamlit = "1.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
